+++ src/components/content/care-team/patient-careteam.tsx
@@ -2,7 +2,6 @@ import cx from "classnames";
 import { useRef } from "react";
 import { patientCareTeamColumns } from "./helpers/columns";
 import { useResourceDetailsDrawer } from "../resource/resource-details-drawer";
-import { useCTW } from "@/components/core/providers/use-ctw";
 import { Table } from "@/components/core/table/table";
 import { usePatientCareTeam } from "@/fhir/care-team";
 import { CareTeamPractitionerModel } from "@/fhir/models/careteam-practitioner";
@@ -24,7 +23,6 @@ export function PatientCareTeam({ className }: PatientCareTeamProps) {
   const { enabled } = useFQSFeatureToggle("careTeams");
   const containerRef = useRef<HTMLDivElement>(null);
   const breakpoints = useBreakpoints(containerRef);
-  const { featureFlags } = useCTW();
   const patientCareTeamQuery = usePatientCareTeam();
 
   const openDetails = useResourceDetailsDrawer({
@@ -50,7 +48,7 @@ export function PatientCareTeam({ className }: PatientCareTeamProps) {
         stacked={breakpoints.sm}
         isLoading={patientCareTeamQuery.isLoading}
         records={patientCareTeamQuery.data ?? []}
-        columns={patientCareTeamColumns(featureFlags?.enableViewFhirButton)}
+        columns={patientCareTeamColumns}
         handleRowClick={openDetails}
       />
     </div>