+++ src/components/content/conditions/unread-conditions-notification.tsx
@@ -1,30 +1,11 @@
-import { useEffect, useState } from "react";
-import { useCTW } from "@/components/core/providers/use-ctw";
 import { UnreadNotification } from "@/components/core/unread-notification";
-import { usePatientConditionsOutside } from "@/services/conditions";
+import { usePatientConditionsAll } from "@/services/conditions";
 
-export const UnreadConditionsNotification = () => {
-  const query = usePatientConditionsOutside();
-  const { getRequestContext } = useCTW();
+export type UnreadConditionsNotificationProps = {
+  className?: string;
+};
 
-  const [userBuilderId, setUserBuilderId] = useState("");
-
-  useEffect(() => {
-    async function load() {
-      const requestContext = await getRequestContext();
-      setUserBuilderId(requestContext.builderId);
-    }
-
-    void load();
-  }, [getRequestContext]);
-
-  const unreadOutsideConditions = query.data.filter(
-    (condition) =>
-      !condition.isDismissed && !condition.isRead && !condition.ownedByBuilder(userBuilderId)
-  );
-
-  if (unreadOutsideConditions.length > 0) {
-    return <UnreadNotification />;
-  }
-  return null;
+export const UnreadConditionsNotification = ({ className }: UnreadConditionsNotificationProps) => {
+  const query = usePatientConditionsAll();
+  return <UnreadNotification className={className} query={query} />;
 };