+++ src/components/content/observations/helpers/base.tsx
@@ -8,6 +8,7 @@ import {
 } from "../../resource/resource-table-actions";
 import { useObservationsDetailsDrawer } from "@/components/content/observations/helpers/drawer";
 import { defaultFilters, filters } from "@/components/content/observations/helpers/filters";
+import { EmptyTable } from "@/components/core/empty-table";
 import { DiagnosticReportModel } from "@/fhir/models";
 import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
 
@@ -15,7 +16,7 @@ export type PatientObservationsTableProps = {
   action?: ResourceTableActionsProps<DiagnosticReportModel>["action"];
   className?: cx.Argument;
   query: { data?: DiagnosticReportModel[]; isLoading: boolean };
-  rowActions?: ResourceTableProps<DiagnosticReportModel>["rowActions"];
+  rowActions?: ResourceTableProps<DiagnosticReportModel>["RowActions"];
 };
 
 export const PatientObservationsBase = ({
@@ -30,6 +31,9 @@ export const PatientObservationsBase = ({
     records: query.data,
     defaultFilters,
   });
+
+  const isEmptyQuery = query.data?.length === 0;
+  const hasZeroFilteredRecords = !isEmptyQuery && data.length === 0;
 
   return (
     <div className={cx(className, "ctw-scrollable-pass-through-height")}>
@@ -51,9 +55,11 @@ export const PatientObservationsBase = ({
         className="ctw-patient-observations"
         columns={diagnosticReportColumns}
         data={data}
-        emptyMessage="There are no observation records available."
+        emptyMessage={
+          <EmptyTable hasZeroFilteredRecords={hasZeroFilteredRecords} resourceName="observations" />
+        }
         isLoading={query.isLoading}
-        rowActions={rowActions}
+        RowActions={rowActions}
         onRowClick={openDrawer}
       />
     </div>