+++ src/components/core/view-fhir.tsx
@@ -6,7 +6,7 @@ import { Drawer } from "./drawer";
 import { useDrawer } from "./providers/drawer-provider";
 
 export type ViewFHIRProps = {
-  name: string;
+  name?: string;
   resource: fhir4.Resource;
 };
 
@@ -20,16 +20,14 @@ export function useFHIRDrawer() {
   };
 }
 
-export function ViewFHIR({ name, resource }: ViewFHIRProps) {
+export function ViewFHIR({ name = "View FHIR", resource }: ViewFHIRProps) {
   const openDrawer = useFHIRDrawer();
 
   return (
     <button
       type="button"
-      className="ctw-btn-primary ctw-text-sm"
-      onClick={(event) => {
-        // Prevents any rowClick handler from firing.
-        event.stopPropagation();
+      className="ctw-btn-default"
+      onClick={() => {
         openDrawer(name, resource);
       }}
     >