+++ src/services/fqs/queries/medication-request.ts
@@ -76,7 +76,36 @@ export const medicationRequestQuery = gql`
           }
           dispenseRequest {
             performer {
-              ...Reference
+              id
+              extension {
+                url
+                valueString
+              }
+              reference
+              type
+              identifier {
+                id
+                system
+                value
+              }
+              display
+              resource {
+                ... on Organization {
+                  id
+                  resourceType
+                  name
+                  telecom {
+                    value
+                  }
+                  address {
+                    city
+                    state
+                    postalCode
+                    text
+                    line
+                  }
+                }
+              }
             }
             numberOfRepeatsAllowed
             initialFill {
@@ -106,7 +135,35 @@ export const medicationRequestQuery = gql`
             ...Reference
           }
           requester {
-            ...Reference
+            id
+            extension {
+              url
+              valueString
+            }
+            reference
+            type
+            identifier {
+              id
+              system
+              value
+            }
+            display
+            resource {
+              ... on Practitioner {
+                id
+                resourceType
+                name {
+                  family
+                  given
+                }
+                address {
+                  text
+                }
+                telecom {
+                  value
+                }
+              }
+            }
           }
           status
         }